pyproject.toml
[project]
name = "vir-bot"
version = "0.1.0"
description = "基于聊天记录蒸馏的AI机器人 — 数字分身"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "hxxim", email = "" }
]
keywords = ["ai", "chatbot", "persona", "digital-twin", "llm", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "aiohttp>=3.10.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "discord.py>=2.4.0",
    "websockets>=12.0",
    "edge-tts>=6.1.0",
    "openai-whisper>=20240930",
    "paho-mqtt>=1.6.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.0.0",
    "pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]